+++ apps/studio/components/interfaces/DataWarehouse/WarehouseMenuItem.tsx
@@ -1,44 +1,32 @@
-import { zodResolver } from '@hookform/resolvers/zod'
 import { PermissionAction } from '@supabase/shared-types/out/constants'
-import { Database, EditIcon, MoreHorizontal, TrashIcon } from 'lucide-react'
-import Link from 'next/link'
+import { Database, EditIcon, TrashIcon } from 'lucide-react'
 import { useRouter } from 'next/router'
 import { useState } from 'react'
-import { useForm } from 'react-hook-form'
 import { toast } from 'sonner'
-import { z } from 'zod'
 
-import { FormMessage } from '@ui/components/shadcn/ui/form'
 import { useParams } from 'common'
 import { useDeleteCollectionMutation } from 'data/analytics/warehouse-collections-delete-mutation'
 import { useUpdateCollection } from 'data/analytics/warehouse-collections-update-mutation'
 import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
 import {
-  Button,
-  DropdownMenu,
-  DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuSeparator,
-  DropdownMenuTrigger,
-  FormControl_Shadcn_,
-  FormField_Shadcn_,
-  Form_Shadcn_,
-  Input_Shadcn_,
   Modal,
   TooltipContent_Shadcn_,
   TooltipTrigger_Shadcn_,
   Tooltip_Shadcn_,
-  cn,
 } from 'ui'
-import { FormItemLayout } from 'ui-patterns/form/FormItemLayout/FormItemLayout'
+
 import { LogsSidebarItem } from '../Settings/Logs/SidebarV2/SidebarItem'
 import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
+import { CollectionForm } from './CollectionForm'
 
 type Props = {
   item: {
     id: number
     token: string
     name: string
+    retention_days: number
   }
 }
 
@@ -71,16 +59,6 @@ export const WarehouseMenuItem = ({ item }: Props) => {
 
   const isLoading = updateCollection.isLoading || deleteCollection.isLoading
 
-  const UpdateFormSchema = z.object({
-    name: z.string().min(1, {
-      message: 'Collection name is required',
-    }),
-  })
-
-  const updateForm = useForm<z.infer<typeof UpdateFormSchema>>({
-    resolver: zodResolver(UpdateFormSchema),
-  })
-
   return (
     <>
       <LogsSidebarItem
@@ -100,12 +78,12 @@ export const WarehouseMenuItem = ({ item }: Props) => {
                   }}
                 >
                   <EditIcon className="mr-2" size={14} />
-                  <div>Rename collection</div>
+                  <div>Update collection</div>
                 </DropdownMenuItem>
               </TooltipTrigger_Shadcn_>
               {!canUpdateCollection && (
                 <TooltipContent_Shadcn_ side="right">
-                  You need additional permissions to rename collections
+                  You need additional permissions to update collections
                 </TooltipContent_Shadcn_>
               )}
             </Tooltip_Shadcn_>
@@ -140,48 +118,19 @@ export const WarehouseMenuItem = ({ item }: Props) => {
         hideFooter
         header={`Update collection "${item.name}"`}
       >
-        <Form_Shadcn_ {...updateForm}>
-          <form
-            id="update-collection-form"
-            onSubmit={updateForm.handleSubmit((data) =>
-              updateCollection.mutate({
-                projectRef,
-                collectionToken: item.token,
-                name: data.name,
-              })
-            )}
-          >
-            <Modal.Content className="space-y-6 py-6">
-              <FormField_Shadcn_
-                control={updateForm.control}
-                name="name"
-                render={({ field }) => (
-                  <FormItemLayout label="Collection name">
-                    <FormControl_Shadcn_>
-                      <Input_Shadcn_ defaultValue={item.name} type="text" {...field} />
-                    </FormControl_Shadcn_>
-                  </FormItemLayout>
-                )}
-              />
-            </Modal.Content>
-            <FormMessage />
-            <div className="flex gap-2 justify-end p-3 border-t">
-              <Button
-                disabled={isLoading}
-                type="outline"
-                onClick={() => {
-                  setShowUpdateDialog(false)
-                }}
-                htmlType="reset"
-              >
-                Cancel
-              </Button>
-              <Button htmlType="submit" loading={isLoading}>
-                Update collection
-              </Button>
-            </div>
-          </form>
-        </Form_Shadcn_>
+        <CollectionForm
+          onCancelClick={() => setShowUpdateDialog(false)}
+          isLoading={isLoading}
+          initialValues={{ name: item.name, retention_days: item.retention_days }}
+          onSubmit={({ name, retention_days }) => {
+            updateCollection.mutate({
+              projectRef,
+              collectionToken: item.token,
+              name,
+              retention_days,
+            })
+          }}
+        />
       </Modal>
       <ConfirmationModal
         variant="destructive"
