+++ apps/studio/data/analytics/warehouse-collections-update-mutation.ts
@@ -13,6 +13,7 @@ export type UpdateCollectionArgs = {
 export type UpdateCollectionPayload = {
   name?: string
   // description?: string
+  retention_days?: number
 }
 
 export async function updateCollection(
@@ -51,7 +52,7 @@ export const useUpdateCollection = ({
     (payload) =>
       updateCollection(
         { projectRef: payload.projectRef, collectionToken: payload.collectionToken },
-        { name: payload.name }
+        { name: payload.name, retention_days: payload.retention_days }
       ),
     {
       async onSuccess(data, variables, context) {
